// web/src/api/models/DhcpStatsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { KeaStat } from './KeaStat';
import {
    KeaStatFromJSON,
    KeaStatFromJSONTyped,
    KeaStatToJSON,
    KeaStatToJSONTyped,
} from './KeaStat';
import type { DhcpServiceState } from './DhcpServiceState';
import {
    DhcpServiceStateFromJSON,
    DhcpServiceStateFromJSONTyped,
    DhcpServiceStateToJSON,
    DhcpServiceStateToJSONTyped,
} from './DhcpServiceState';

/**
 * 
 * @export
 * @interface DhcpStatsResponse
 */
export interface DhcpStatsResponse {
    /**
     * 
     * @type {Array<DhcpServiceState>}
     * @memberof DhcpStatsResponse
     */
    services?: Array<DhcpServiceState>;
    /**
     * 
     * @type {Array<KeaStat>}
     * @memberof DhcpStatsResponse
     */
    stats4?: Array<KeaStat>;
    /**
     * 
     * @type {Array<KeaStat>}
     * @memberof DhcpStatsResponse
     */
    stats6?: Array<KeaStat>;
}

/**
 * Check if a given object implements the DhcpStatsResponse interface.
 */
export function instanceOfDhcpStatsResponse(value: object): value is DhcpStatsResponse {
    return true;
}

export function DhcpStatsResponseFromJSON(json: any): DhcpStatsResponse {
    return DhcpStatsResponseFromJSONTyped(json, false);
}

export function DhcpStatsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DhcpStatsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'services': json['services'] == null ? undefined : ((json['services'] as Array<any>).map(DhcpServiceStateFromJSON)),
        'stats4': json['stats4'] == null ? undefined : ((json['stats4'] as Array<any>).map(KeaStatFromJSON)),
        'stats6': json['stats6'] == null ? undefined : ((json['stats6'] as Array<any>).map(KeaStatFromJSON)),
    };
}

export function DhcpStatsResponseToJSON(json: any): DhcpStatsResponse {
    return DhcpStatsResponseToJSONTyped(json, false);
}

export function DhcpStatsResponseToJSONTyped(value?: DhcpStatsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'services': value['services'] == null ? undefined : ((value['services'] as Array<any>).map(DhcpServiceStateToJSON)),
        'stats4': value['stats4'] == null ? undefined : ((value['stats4'] as Array<any>).map(KeaStatToJSON)),
        'stats6': value['stats6'] == null ? undefined : ((value['stats6'] as Array<any>).map(KeaStatToJSON)),
    };
}