// web/src/api/models/KeaSubnet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface KeaSubnet
 */
export interface KeaSubnet {
    /**
     * 
     * @type {number}
     * @memberof KeaSubnet
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof KeaSubnet
     */
    subnet: string;
}

/**
 * Check if a given object implements the KeaSubnet interface.
 */
export function instanceOfKeaSubnet(value: object): value is KeaSubnet {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('subnet' in value) || value['subnet'] === undefined) return false;
    return true;
}

export function KeaSubnetFromJSON(json: any): KeaSubnet {
    return KeaSubnetFromJSONTyped(json, false);
}

export function KeaSubnetFromJSONTyped(json: any, ignoreDiscriminator: boolean): KeaSubnet {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'subnet': json['subnet'],
    };
}

export function KeaSubnetToJSON(json: any): KeaSubnet {
    return KeaSubnetToJSONTyped(json, false);
}

export function KeaSubnetToJSONTyped(value?: KeaSubnet | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'subnet': value['subnet'],
    };
}